expose create_and_serialize_proof + fix error in binaries
