+++ stwo_cairo_prover/Cargo.toml
@@ -67,6 +67,7 @@ serde_arrays = "0.1.0"
 rayon = "1.10.0"
 indoc = "2.0.6"
 tempfile = "3.8.1"
+anyhow = "1.0.98"
 
 [profile.bench]
 codegen-units = 1
