+++ stwo_cairo_prover/crates/dev_utils/src/utils.rs
@@ -1,9 +1,6 @@
 use std::fs::File;
 use std::path::PathBuf;
 
-use cairo_air::utils::{serialize_proof_to_file, ProofFormat};
-use cairo_air::verifier::{verify_cairo, CairoVerificationError};
-use cairo_air::PreProcessedTraceVariant;
 use cairo_lang_executable::executable::{EntryPointKind, Executable};
 use cairo_lang_runner::{build_hints_dict, Arg, CairoHintProcessor};
 use cairo_lang_utils::bigint::BigUintAsHex;
@@ -15,108 +12,9 @@ use cairo_vm::types::program::Program;
 use cairo_vm::types::relocatable::MaybeRelocatable;
 use cairo_vm::vm::runners::cairo_runner::CairoRunner;
 use cairo_vm::Felt252;
-use serde::Serialize;
 use serde_json::from_reader;
-use stwo::core::channel::MerkleChannel;
-use stwo::core::pcs::PcsConfig;
-use stwo::core::vcs::blake2_merkle::Blake2sMerkleChannel;
-use stwo::core::vcs::poseidon252_merkle::Poseidon252MerkleChannel;
-use stwo::core::vcs::MerkleHasher;
-use stwo::prover::backend::simd::SimdBackend;
-use stwo::prover::backend::BackendForChannel;
-use stwo::prover::ProvingError;
 use stwo_cairo_adapter::adapter::adapter;
-use stwo_cairo_adapter::vm_import::VmImportError;
 use stwo_cairo_adapter::ProverInput;
-use stwo_cairo_prover::prover::{
-    default_prod_prover_parameters, prove_cairo, ChannelHash, ProverParameters,
-};
-use stwo_cairo_serialize::CairoSerialize;
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum Error {
-    #[error("Invalid arguments: {0}")]
-    Cli(#[from] clap::Error),
-    #[error("IO failed: {0}")]
-    IO(#[from] std::io::Error),
-    #[error("Proving failed: {0}")]
-    Proving(#[from] ProvingError),
-    #[error("Serialization failed: {0}")]
-    Serializing(#[from] sonic_rs::error::Error),
-    #[error("Verification failed: {0}")]
-    Verification(#[from] CairoVerificationError),
-    #[error("VM import failed: {0}")]
-    VmImport(#[from] VmImportError),
-}
-
-/// Generates proof given the Cairo VM output and prover config/parameters.
-/// Serializes the proof as JSON and write to the output path.
-/// Verifies the proof in case the respective flag is set.
-pub fn create_and_serialize_generic_proof<MC: MerkleChannel>(
-    input: ProverInput,
-    pcs_config: PcsConfig,
-    preprocessed_trace: PreProcessedTraceVariant,
-    verify: bool,
-    proof_path: PathBuf,
-    proof_format: ProofFormat,
-) -> Result<(), Error>
-where
-    SimdBackend: BackendForChannel<MC>,
-    MC::H: Serialize,
-    <MC::H as MerkleHasher>::Hash: CairoSerialize,
-{
-    let proof = prove_cairo::<MC>(input, pcs_config, preprocessed_trace)?;
-
-    serialize_proof_to_file::<MC::H>(&proof, &proof_path, proof_format)?;
-
-    if verify {
-        verify_cairo::<MC>(proof, preprocessed_trace)?;
-        log::info!("Proof verified successfully");
-    }
-
-    Ok(())
-}
-
-pub fn create_and_serialize_proof(
-    input: ProverInput,
-    verify: bool,
-    proof_path: PathBuf,
-    proof_format: ProofFormat,
-    proof_params_json: Option<PathBuf>,
-) -> Result<(), Error> {
-    let ProverParameters {
-        channel_hash,
-        pcs_config,
-        preprocessed_trace,
-    } = match proof_params_json {
-        Some(path) => sonic_rs::from_str(&std::fs::read_to_string(&path)?)?,
-        None => default_prod_prover_parameters(),
-    };
-
-    let create_and_serialize_generic_proof: fn(
-        ProverInput,
-        PcsConfig,
-        PreProcessedTraceVariant,
-        bool,
-        PathBuf,
-        ProofFormat,
-    ) -> Result<(), Error> = match channel_hash {
-        ChannelHash::Blake2s => create_and_serialize_generic_proof::<Blake2sMerkleChannel>,
-        ChannelHash::Poseidon252 => create_and_serialize_generic_proof::<Poseidon252MerkleChannel>,
-    };
-
-    create_and_serialize_generic_proof(
-        input,
-        pcs_config,
-        preprocessed_trace,
-        verify,
-        proof_path,
-        proof_format,
-    )?;
-
-    Ok(())
-}
 
 pub fn runner_from_compiled_program(
     program_path: &PathBuf,
