+++ stwo_cairo_prover/crates/prover/Cargo.toml
@@ -25,6 +25,8 @@ stwo-cairo-adapter.workspace = true
 rayon.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+sonic-rs.workspace = true
+anyhow.workspace = true
 starknet-curve = "0.5.1"
 starknet-ff.workspace = true
 starknet-types-core.workspace = true
